# docs/swagger.json
{
    "schemes": ["http"],
    "swagger": "2.0",
    "info": {
        "description": "API Kasir dengan Layered Architecture - Week 2 Challenge",
        "title": "Kasir API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@kodingworks.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/categories": {
            "get": {
                "description": "Get all categories",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["categories"],
                "summary": "List all categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Category"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new category",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["categories"],
                "summary": "Create category",
                "parameters": [
                    {
                        "description": "Category data",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Category"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Category"
                        }
                    }
                }
            }
        },
...
        "/api/produk": {
            "get": {
                "description": "Get all products",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["products"],
                "summary": "List all products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Product"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new product",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["products"],
                "summary": "Create product",
                "parameters": [
                    {
                        "description": "Product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Product"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "entity.Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "harga": {
                    "type": "integer"
                },
                "category_id": {
                    "type": "integer"
                },
                "category": {
                    "$ref": "#/definitions/entity.Category"
                }
            }
        }
    }
}